requirements.txt
apiview==1.3.25
APScheduler==3.6.3
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-apscheduler==0.3.0
django-import-export==2.0.1
django-rest-auth==0.9.5
django-suit==0.2.25
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
tzlocal==2.0.0
xlrd==1.2.0
xlwt==1.3.0
